requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.4
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.5
dataclasses-json==0.6.7
distro==1.9.0
faiss-cpu==1.11.0.post1
filelock==3.18.0
frozenlist==1.7.0
fsspec==2025.7.0
groq==0.30.0
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.3
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.72
langchain-groq==0.3.6
langchain-text-splitters==0.3.9
langsmith==0.4.8
MarkupSafe==3.0.2
marshmallow==3.26.1
mpmath==1.3.0
multidict==6.6.3
mypy_extensions==1.1.0
networkx==3.5
numpy==2.3.2
orjson==3.11.1
packaging==25.0
pdfminer.six==20250506
pillow==11.3.0
python-dotenv==1.1.1
PyYAML==6.0.2
regex==2025.7.31
requests==2.32.4
requests-toolbelt==1.0.0
safetensors==0.5.3
scikit-learn==1.7.1
scipy==1.16.1
sentence-transformers==5.0.0
setuptools==80.9.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.42
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.21.4
torch==2.7.1
tqdm==4.67.1
transformers==4.54.1
typing_extensions==4.14.1
typing-inspect==0.9.0
typing-inspection==0.4.1
urllib3==2.5.0
wikipedia==1.4.0
yarl==1.20.1
zstandard==0.23.0
